+++ src/main/java/kitchenpos/domain/menu/Menu.java
@@ -1,4 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.domain.menu;
+
+import kitchenpos.domain.menugroup.MenuGroup;
 
 import javax.persistence.Embedded;
 import javax.persistence.Entity;
