+++ src/main/java/kitchenpos/domain/menu/MenuPrice.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.menu;
 
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
