+++ src/main/java/kitchenpos/domain/menugroup/MenuGroup.java
@@ -1,32 +1,33 @@
-package kitchenpos.domain;
+package kitchenpos.domain.menugroup;
 
+import javax.persistence.Embedded;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.validation.constraints.NotNull;
-import java.time.LocalDateTime;
 
 @Entity
-public class TableGroup {
+public class MenuGroup {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     @NotNull
-    private LocalDateTime createdDate;
+    @Embedded
+    private MenuGroupName name;
 
-    public TableGroup() {
+    public MenuGroup() {
     }
 
-    public TableGroup(final LocalDateTime localDateTime) {
-        this.createdDate = localDateTime;
+    public MenuGroup(final MenuGroupName name) {
+        this.name = name;
     }
 
     public Long getId() {
         return id;
     }
 
-    public LocalDateTime getCreatedDate() {
-        return createdDate;
+    public String getName() {
+        return name.getName();
     }
 }
