+++ src/main/java/kitchenpos/domain/menuproduct/MenuProduct.java
@@ -1,34 +1,40 @@
-package kitchenpos.domain;
+package kitchenpos.domain.menuproduct;
+
+import kitchenpos.domain.menu.Menu;
+import kitchenpos.domain.product.Product;
 
 import javax.persistence.Embedded;
 import javax.persistence.Entity;
+import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.OneToOne;
+import javax.validation.constraints.NotNull;
 
 @Entity
-public class OrderLineItem {
+public class MenuProduct {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long seq;
-    @ManyToOne
-    @JoinColumn(name = "order_id")
-    private Order order;
-    @OneToOne
+    @NotNull
+    @Embedded
+    private MenuProductQuantity quantity;
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "menu_id")
     private Menu menu;
-    @Embedded
-    private OrderLineItemQuantity quantity;
+    @OneToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "product_id")
+    private Product product;
 
-    public OrderLineItem() {
+    public MenuProduct() {
     }
 
-    public OrderLineItem(final Order order, final Menu menu, final OrderLineItemQuantity quantity) {
-        this.order = order;
+    public MenuProduct(final Menu menu, final Product product, final MenuProductQuantity quantity) {
         this.menu = menu;
+        this.product = product;
         this.quantity = quantity;
     }
 
@@ -36,12 +42,12 @@ public class OrderLineItem {
         return seq;
     }
 
-    public Order getOrder() {
-        return order;
-    }
-
     public Menu getMenu() {
         return menu;
+    }
+
+    public Product getProduct() {
+        return product;
     }
 
     public long getQuantity() {
