+++ src/main/java/kitchenpos/domain/order/Order.java
@@ -1,4 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.domain.order;
+
+import kitchenpos.domain.ordertable.OrderTable;
 
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
