+++ src/main/java/kitchenpos/domain/orderlineitem/OrderLineItem.java
@@ -1,37 +1,37 @@
-package kitchenpos.domain;
+package kitchenpos.domain.orderlineitem;
+
+import kitchenpos.domain.menu.Menu;
+import kitchenpos.domain.order.Order;
 
 import javax.persistence.Embedded;
 import javax.persistence.Entity;
-import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.OneToOne;
-import javax.validation.constraints.NotNull;
 
 @Entity
-public class MenuProduct {
+public class OrderLineItem {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long seq;
-    @NotNull
-    @Embedded
-    private MenuProductQuantity quantity;
-    @ManyToOne(fetch = FetchType.LAZY)
+    @ManyToOne
+    @JoinColumn(name = "order_id")
+    private Order order;
+    @OneToOne
     @JoinColumn(name = "menu_id")
     private Menu menu;
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "product_id")
-    private Product product;
+    @Embedded
+    private OrderLineItemQuantity quantity;
 
-    public MenuProduct() {
+    public OrderLineItem() {
     }
 
-    public MenuProduct(final Menu menu, final Product product, final MenuProductQuantity quantity) {
+    public OrderLineItem(final Order order, final Menu menu, final OrderLineItemQuantity quantity) {
+        this.order = order;
         this.menu = menu;
-        this.product = product;
         this.quantity = quantity;
     }
 
@@ -39,12 +39,12 @@ public class MenuProduct {
         return seq;
     }
 
-    public Menu getMenu() {
-        return menu;
+    public Order getOrder() {
+        return order;
     }
 
-    public Product getProduct() {
-        return product;
+    public Menu getMenu() {
+        return menu;
     }
 
     public long getQuantity() {
