+++ src/main/java/kitchenpos/domain/ordertable/OrderTableNumberOfGuests.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.ordertable;
 
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
