+++ src/main/java/kitchenpos/domain/product/ProductPrice.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.product;
 
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
