+++ src/main/java/kitchenpos/domain/tablegroup/TableGroup.java
@@ -1,33 +1,32 @@
-package kitchenpos.domain;
+package kitchenpos.domain.tablegroup;
 
-import javax.persistence.Embedded;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.validation.constraints.NotNull;
+import java.time.LocalDateTime;
 
 @Entity
-public class MenuGroup {
+public class TableGroup {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     @NotNull
-    @Embedded
-    private MenuGroupName name;
+    private LocalDateTime createdDate;
 
-    public MenuGroup() {
+    public TableGroup() {
     }
 
-    public MenuGroup(final MenuGroupName name) {
-        this.name = name;
+    public TableGroup(final LocalDateTime localDateTime) {
+        this.createdDate = localDateTime;
     }
 
     public Long getId() {
         return id;
     }
 
-    public String getName() {
-        return name.getName();
+    public LocalDateTime getCreatedDate() {
+        return createdDate;
     }
 }
