+++ src/main/java/kitchenpos/repository/OrderRepository.java
@@ -1,8 +1,8 @@
 package kitchenpos.repository;
 
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderStatus;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.order.Order;
+import kitchenpos.domain.order.OrderStatus;
+import kitchenpos.domain.ordertable.OrderTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
