+++ src/main/java/kitchenpos/repository/OrderTableRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.repository;
 
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.domain.ordertable.OrderTable;
+import kitchenpos.domain.tablegroup.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
